requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
numpy==1.26.0
obsws-python==1.6.0
openai==0.28.1
pycparser==2.21
pytchat==0.5.5
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
tqdm==4.66.1
urllib3==2.0.4
websocket-client==1.6.3
yarl==1.9.2
